requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0,<4.0.0
requests>=2.25.0,<3.0.0
gunicorn>=20.1.0,<21.0.0
django-cors-headers>=3.10.0
